Normalized formatting for all code files.

+++ branch-0.90x/fCraft/Utils/ZipStorer.cs
@@ -1,5 +1,6 @@
 ﻿// Part of fCraft | Copyright 2009-2013 Matvei Stefarov <me@matvei.org> | BSD-3 | See LICENSE.txt
 // Based on ZipStorer 2.35 (March 14, 2010), by Jaime Olivares, zipstorer.codeplex.com
+
 using System.Collections.Generic;
 using System.Text;
 using fCraft;
@@ -60,7 +61,6 @@ namespace System.IO.Compression {
             }
         }
 
-
         #region Public properties
 
         /// <summary> True if UTF8 encoding for filename and comments, false if default (CP 437). </summary>
@@ -70,7 +70,6 @@ namespace System.IO.Compression {
         public bool ForceDeflating { get; set; }
 
         #endregion
-
 
         #region Private fields
 
@@ -95,7 +94,6 @@ namespace System.IO.Compression {
 
         #endregion
 
-
         #region Public methods
 
         // Static constructor. Just invoked once in order to create the CRC32 lookup table.
@@ -105,10 +103,8 @@ namespace System.IO.Compression {
             for( int i = 0; i < CrcTable.Length; i++ ) {
                 UInt32 c = (UInt32)i;
                 for( int j = 0; j < 8; j++ ) {
-                    if( ( c & 1 ) != 0 )
-                        c = 3988292384 ^ ( c >> 1 );
-                    else
-                        c >>= 1;
+                    if( (c & 1) != 0 ) c = 3988292384 ^ (c >> 1);
+                    else c >>= 1;
                 }
                 CrcTable[i] = c;
             }
@@ -173,16 +169,14 @@ namespace System.IO.Compression {
         [NotNull]
         public static ZipStorer Open( [NotNull] Stream stream, FileAccess fileFileAccess ) {
             if( stream == null ) throw new ArgumentNullException( "stream" );
-            if( !stream.CanSeek && fileFileAccess != FileAccess.Read )
-                throw new InvalidOperationException( "Stream cannot seek" );
+            if( !stream.CanSeek && fileFileAccess != FileAccess.Read ) throw new InvalidOperationException( "Stream cannot seek" );
 
             ZipStorer zip = new ZipStorer {
                 zipFileStream = stream,
                 access = fileFileAccess
             };
 
-            if( zip.ReadFileInfo() )
-                return zip;
+            if( zip.ReadFileInfo() ) return zip;
 
             throw new InvalidDataException();
         }
@@ -213,15 +207,14 @@ namespace System.IO.Compression {
                                DateTime modTime, [CanBeNull] string fileComment ) {
             if( fileNameInZip == null ) throw new ArgumentNullException( "fileNameInZip" );
             if( source == null ) throw new ArgumentNullException( "source" );
-            if( access == FileAccess.Read )
-                throw new InvalidOperationException( "Writing is not allowed" );
+            if( access == FileAccess.Read ) throw new InvalidOperationException( "Writing is not allowed" );
 
             // Prepare the fileinfo
             ZipFileEntry zfe = new ZipFileEntry {
                 Method = method,
                 EncodeUTF8 = EncodeUTF8,
                 FileNameInZip = NormalizedFileName( fileNameInZip ),
-                Comment = ( fileComment ?? "" ),
+                Comment = (fileComment ?? ""),
                 Crc32 = 0,
                 HeaderOffset = (uint)zipFileStream.Position,
                 ModifyTime = modTime
@@ -250,19 +243,16 @@ namespace System.IO.Compression {
                 uint centralOffset = (uint)zipFileStream.Position;
                 uint centralSize = 0;
 
-                if( centralDirImage != null )
-                    zipFileStream.Write( centralDirImage, 0, centralDirImage.Length );
+                if( centralDirImage != null ) zipFileStream.Write( centralDirImage, 0, centralDirImage.Length );
 
                 for( int i = 0; i < files.Count; i++ ) {
                     long pos = zipFileStream.Position;
                     WriteCentralDirRecord( files[i] );
-                    centralSize += (uint)( zipFileStream.Position - pos );
+                    centralSize += (uint)(zipFileStream.Position - pos);
                 }
 
-                if( centralDirImage != null )
-                    WriteEndRecord( centralSize + (uint)centralDirImage.Length, centralOffset );
-                else
-                    WriteEndRecord( centralSize, centralOffset );
+                if( centralDirImage != null ) WriteEndRecord( centralSize + (uint)centralDirImage.Length, centralOffset );
+                else WriteEndRecord( centralSize, centralOffset );
             }
 
             if( zipFileStream == null ) return;
@@ -277,17 +267,15 @@ namespace System.IO.Compression {
         /// <returns> List of all entries in directory. </returns>
         [NotNull]
         public List<ZipFileEntry> ReadCentralDir() {
-            if( centralDirImage == null )
-                throw new InvalidOperationException( "Central directory currently does not exist" );
+            if( centralDirImage == null ) throw new InvalidOperationException( "Central directory currently does not exist" );
 
             List<ZipFileEntry> result = new List<ZipFileEntry>();
 
             for( int pointer = 0; pointer < centralDirImage.Length; ) {
                 uint signature = BitConverter.ToUInt32( centralDirImage, pointer );
-                if( signature != 0x02014b50 )
-                    break;
+                if( signature != 0x02014b50 ) break;
 
-                bool encodeUTF8 = ( BitConverter.ToUInt16( centralDirImage, pointer + 8 ) & 0x0800 ) != 0;
+                bool encodeUTF8 = (BitConverter.ToUInt16( centralDirImage, pointer + 8 ) & 0x0800) != 0;
                 ushort method = BitConverter.ToUInt16( centralDirImage, pointer + 10 );
                 uint modifyTime = BitConverter.ToUInt32( centralDirImage, pointer + 12 );
                 uint crc32 = BitConverter.ToUInt32( centralDirImage, pointer + 16 );
@@ -297,7 +285,7 @@ namespace System.IO.Compression {
                 ushort extraSize = BitConverter.ToUInt16( centralDirImage, pointer + 30 );
                 ushort commentSize = BitConverter.ToUInt16( centralDirImage, pointer + 32 );
                 uint headerOffset = BitConverter.ToUInt32( centralDirImage, pointer + 42 );
-                uint headerSize = (uint)( 46 + filenameSize + extraSize + commentSize );
+                uint headerSize = (uint)(46 + filenameSize + extraSize + commentSize);
 
                 Encoding encoder = encodeUTF8 ? Encoding.UTF8 : DefaultEncoding;
 
@@ -321,7 +309,7 @@ namespace System.IO.Compression {
                                                      commentSize );
 
                 result.Add( zfe );
-                pointer += ( 46 + filenameSize + extraSize + commentSize );
+                pointer += (46 + filenameSize + extraSize + commentSize);
             }
 
             return result;
@@ -337,16 +325,13 @@ namespace System.IO.Compression {
             // Make sure the parent directory exist
             string path = Paths.GetDirectoryNameOrRoot( filename );
 
-            if( !Directory.Exists( path ) )
-                Directory.CreateDirectory( path );
+            if( !Directory.Exists( path ) ) Directory.CreateDirectory( path );
             // Check it is directory. If so, do nothing
-            if( Directory.Exists( filename ) )
-                return true;
+            if( Directory.Exists( filename ) ) return true;
 
             Stream output = new FileStream( filename, FileMode.Create, FileAccess.Write );
             bool result = ExtractFile( zfe, output );
-            if( result )
-                output.Close();
+            if( result ) output.Close();
 
             File.SetCreationTime( filename, zfe.ModifyTime );
             File.SetLastWriteTime( filename, zfe.ModifyTime );
@@ -362,15 +347,13 @@ namespace System.IO.Compression {
         /// <remarks> Unique compression methods are Store and Deflate. </remarks>
         public bool ExtractFile( ZipFileEntry zfe, [NotNull] Stream stream ) {
             if( stream == null ) throw new ArgumentNullException( "stream" );
-            if( !stream.CanWrite )
-                throw new InvalidOperationException( "Stream cannot be written" );
+            if( !stream.CanWrite ) throw new InvalidOperationException( "Stream cannot be written" );
 
             // check signature
             byte[] signature = new byte[4];
             zipFileStream.Seek( zfe.HeaderOffset, SeekOrigin.Begin );
             zipFileStream.Read( signature, 0, 4 );
-            if( BitConverter.ToUInt32( signature, 0 ) != 0x04034b50 )
-                return false;
+            if( BitConverter.ToUInt32( signature, 0 ) != 0x04034b50 ) return false;
 
             // Select input stream for inflating or just reading
             Stream inStream;
@@ -396,8 +379,7 @@ namespace System.IO.Compression {
             }
             stream.Flush();
 
-            if( zfe.Method == Compression.Deflate )
-                inStream.Dispose();
+            if( zfe.Method == Compression.Deflate ) inStream.Dispose();
             return true;
         }
 
@@ -410,8 +392,7 @@ namespace System.IO.Compression {
         public static bool RemoveEntries( [NotNull] ref ZipStorer zip, [NotNull] List<ZipFileEntry> zfes ) {
             if( zip == null ) throw new ArgumentNullException( "zip" );
             if( zfes == null ) throw new ArgumentNullException( "zfes" );
-            if( !( zip.zipFileStream is FileStream ) )
-                throw new InvalidOperationException( "RemoveEntries is allowed just over streams of type FileStream" );
+            if( !(zip.zipFileStream is FileStream) ) throw new InvalidOperationException( "RemoveEntries is allowed just over streams of type FileStream" );
 
 
             //Get full list of entries
@@ -444,16 +425,13 @@ namespace System.IO.Compression {
             } catch {
                 return false;
             } finally {
-                if( File.Exists( tempZipName ) )
-                    File.Delete( tempZipName );
-                if( File.Exists( tempEntryName ) )
-                    File.Delete( tempEntryName );
+                if( File.Exists( tempZipName ) ) File.Delete( tempZipName );
+                if( File.Exists( tempEntryName ) ) File.Delete( tempEntryName );
             }
             return true;
         }
 
         #endregion
-
 
         #region Private methods
 
@@ -467,7 +445,7 @@ namespace System.IO.Compression {
             zipFileStream.Read( buffer, 0, 2 );
             ushort extraSize = BitConverter.ToUInt16( buffer, 0 );
 
-            return (uint)( 30 + filenameSize + extraSize + headerOffset );
+            return (uint)(30 + filenameSize + extraSize + headerOffset);
         }
 
 
@@ -494,19 +472,19 @@ namespace System.IO.Compression {
             Encoding encoder = zfe.EncodeUTF8 ? Encoding.UTF8 : DefaultEncoding;
             byte[] encodedFileName = encoder.GetBytes( zfe.FileNameInZip );
 
-            zipFileStream.Write( new byte[] { 80, 75, 3, 4, 20, 0 }, 0, 6 ); // No extra header
-            zipFileStream.Write( BitConverter.GetBytes( (ushort)( zfe.EncodeUTF8 ? 0x0800 : 0 ) ), 0, 2 );
+            zipFileStream.Write( new byte[] {80, 75, 3, 4, 20, 0}, 0, 6 ); // No extra header
+            zipFileStream.Write( BitConverter.GetBytes( (ushort)(zfe.EncodeUTF8 ? 0x0800 : 0) ), 0, 2 );
             // filename and comment encoding 
             zipFileStream.Write( BitConverter.GetBytes( (ushort)zfe.Method ), 0, 2 ); // zipping method
             zipFileStream.Write( BitConverter.GetBytes( DateTimeToDosTime( zfe.ModifyTime ) ), 0, 4 );
             // zipping date and time
-            zipFileStream.Write( new byte[] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 }, 0, 12 );
+            zipFileStream.Write( new byte[] {0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0}, 0, 12 );
             // unused CRC, un/compressed size, updated later
             zipFileStream.Write( BitConverter.GetBytes( (ushort)encodedFileName.Length ), 0, 2 ); // filename length
             zipFileStream.Write( BitConverter.GetBytes( (ushort)0 ), 0, 2 ); // extra length
 
             zipFileStream.Write( encodedFileName, 0, encodedFileName.Length );
-            zfe.HeaderSize = (uint)( zipFileStream.Position - pos );
+            zfe.HeaderSize = (uint)(zipFileStream.Position - pos);
         }
 
 
@@ -540,8 +518,8 @@ namespace System.IO.Compression {
             byte[] encodedFileName = encoder.GetBytes( zfe.FileNameInZip );
             byte[] encodedComment = encoder.GetBytes( zfe.Comment );
 
-            zipFileStream.Write( new byte[] { 80, 75, 1, 2, 23, 0xB, 20, 0 }, 0, 8 );
-            zipFileStream.Write( BitConverter.GetBytes( (ushort)( zfe.EncodeUTF8 ? 0x0800 : 0 ) ), 0, 2 );
+            zipFileStream.Write( new byte[] {80, 75, 1, 2, 23, 0xB, 20, 0}, 0, 8 );
+            zipFileStream.Write( BitConverter.GetBytes( (ushort)(zfe.EncodeUTF8 ? 0x0800 : 0) ), 0, 2 );
             // filename and comment encoding 
             zipFileStream.Write( BitConverter.GetBytes( (ushort)zfe.Method ), 0, 2 ); // zipping method
             zipFileStream.Write( BitConverter.GetBytes( DateTimeToDosTime( zfe.ModifyTime ) ), 0, 4 );
@@ -587,7 +565,7 @@ namespace System.IO.Compression {
             Encoding encoder = EncodeUTF8 ? Encoding.UTF8 : DefaultEncoding;
             byte[] encodedComment = encoder.GetBytes( comment );
 
-            zipFileStream.Write( new byte[] { 80, 75, 5, 6, 0, 0, 0, 0 }, 0, 8 );
+            zipFileStream.Write( new byte[] {80, 75, 5, 6, 0, 0, 0, 0}, 0, 8 );
             zipFileStream.Write( BitConverter.GetBytes( (ushort)files.Count + existingFiles ), 0, 2 );
             zipFileStream.Write( BitConverter.GetBytes( (ushort)files.Count + existingFiles ), 0, 2 );
             zipFileStream.Write( BitConverter.GetBytes( size ), 0, 4 );
@@ -624,7 +602,7 @@ namespace System.IO.Compression {
                         outStream.Write( buffer, 0, bytesRead );
 
                         for( uint i = 0; i < bytesRead; i++ ) {
-                            zfe.Crc32 = CrcTable[( zfe.Crc32 ^ buffer[i] ) & 0xFF] ^ ( zfe.Crc32 >> 8 );
+                            zfe.Crc32 = CrcTable[(zfe.Crc32 ^ buffer[i]) & 0xFF] ^ (zfe.Crc32 >> 8);
                         }
                     }
                 } while( bytesRead == buffer.Length );
@@ -639,10 +617,11 @@ namespace System.IO.Compression {
 
                 zfe.Crc32 ^= 0xffffffff;
                 zfe.FileSize = totalRead;
-                zfe.CompressedSize = (uint)( zipFileStream.Position - posStart );
+                zfe.CompressedSize = (uint)(zipFileStream.Position - posStart);
 
                 // Verify for real compression
-                if( zfe.Method == Compression.Deflate && !ForceDeflating && source.CanSeek && zfe.CompressedSize > zfe.FileSize ) {
+                if( zfe.Method == Compression.Deflate && !ForceDeflating && source.CanSeek &&
+                    zfe.CompressedSize > zfe.FileSize ) {
                     // Start operation again with Store algorithm
                     zfe.Method = Compression.Store;
                     zipFileStream.Position = posStart;
@@ -669,19 +648,19 @@ namespace System.IO.Compression {
 
         static uint DateTimeToDosTime( DateTime dt ) {
             return (uint)(
-                             ( dt.Second / 2 ) | ( dt.Minute << 5 ) | ( dt.Hour << 11 ) |
-                             ( dt.Day << 16 ) | ( dt.Month << 21 ) | ( ( dt.Year - 1980 ) << 25 ) );
+                             (dt.Second/2) | (dt.Minute << 5) | (dt.Hour << 11) |
+                             (dt.Day << 16) | (dt.Month << 21) | ((dt.Year - 1980) << 25));
         }
 
 
         static DateTime DosTimeToDateTime( uint dt ) {
             return new DateTime(
-                (int)( dt >> 25 ) + 1980,
-                (int)( dt >> 21 ) & 15,
-                (int)( dt >> 16 ) & 31,
-                (int)( dt >> 11 ) & 31,
-                (int)( dt >> 5 ) & 63,
-                (int)( dt & 31 ) * 2 );
+                (int)(dt >> 25) + 1980,
+                (int)(dt >> 21) & 15,
+                (int)(dt >> 16) & 31,
+                (int)(dt >> 11) & 31,
+                (int)(dt >> 5) & 63,
+                (int)(dt & 31)*2 );
         }
 
 
@@ -721,8 +700,7 @@ namespace System.IO.Compression {
             string normalizedFileName = fileName.Replace( '\\', '/' );
 
             int pos = normalizedFileName.IndexOf( ':' );
-            if( pos >= 0 )
-                normalizedFileName = normalizedFileName.Remove( 0, pos + 1 );
+            if( pos >= 0 ) normalizedFileName = normalizedFileName.Remove( 0, pos + 1 );
 
             return normalizedFileName.Trim( '/' );
         }
@@ -730,8 +708,7 @@ namespace System.IO.Compression {
 
         // Reads the end-of-central-directory record
         bool ReadFileInfo() {
-            if( zipFileStream.Length < 22 )
-                return false;
+            if( zipFileStream.Length < 22 ) return false;
 
             try {
                 zipFileStream.Seek( -17, SeekOrigin.End );
@@ -749,8 +726,7 @@ namespace System.IO.Compression {
                     UInt16 commentSize = br.ReadUInt16();
 
                     // check if comment field is the very last data in file
-                    if( zipFileStream.Position + commentSize != zipFileStream.Length )
-                        return false;
+                    if( zipFileStream.Position + commentSize != zipFileStream.Length ) return false;
 
                     // Copy entire central directory to a memory buffer
                     existingFiles = entries;
@@ -762,13 +738,12 @@ namespace System.IO.Compression {
                     zipFileStream.Seek( centralDirOffset, SeekOrigin.Begin );
                     return true;
                 } while( zipFileStream.Position > 0 );
-            } catch { }
+            } catch {}
 
             return false;
         }
 
         #endregion
-
 
         #region IDisposable Members
